+++ Changelog.md
@@ -5,6 +5,7 @@
 ### Features and Improvements
 
 - Add type selection, including inline JSON validation, to the Secrets page [#370](https://github.com/PrefectHQ/ui/pull/370)
+- Add an info card about Vault and secret management to the Secrets page [#377](https://github.com/PrefectHQ/ui/pull/377)
 
 ### Bugfixes